\name{mvSLOUCH-internal}
\title{Internal mvSLOUCH objects}
\alias{.beginEstimationParams}
\alias{.bm.phyl.cov}
\alias{.bm.phyl.mean}
\alias{.calc.cov.ouch.mv}
\alias{.calc.cov.slouch.mv}
\alias{.calc.exptA}
\alias{.calc.exptA.2dim}
\alias{.calc.integral.evAStevA}
\alias{.calc.integral.evAStevA.2dim}
\alias{.calc.intercept}
\alias{.calc.mean.ouch.mv}
\alias{.calc.mean.slouch.mv}
\alias{.calc.phyl.cov}
\alias{.calc.phyl.halflife}
\alias{.calc.phyl.LogLik.traits}
\alias{.calc.phyl.mean}
\alias{.calc.vec.dist}
\alias{.calcAncestorTimes}
\alias{.calcAprecalcMatrices}
\alias{.calcCI}
\alias{.calcLogLikParametrized}
\alias{.calcTreeDist}
\alias{.calculate.Tree.dists}
\alias{.CalcVlq}
\alias{.CalcVlq2}
\alias{.CalcVlqStat}
\alias{.ci.HL.loglikfunc}
\alias{.ci.listParams}
\alias{.ci.loglikfunc}
\alias{.ci.vectorizeParams}
\alias{.cleanUpModelParams}
\alias{.col.rot}
\alias{.correct.names}
\alias{.decompEigenA.S}
\alias{.drawRandomNiches}
\alias{.drawRandomRegimes}
\alias{.dzeta.matrix.t}
\alias{.dzeta.matrix.t1t2}
\alias{.dzeta.matrix.t1t2v2}
\alias{.dzeta.matrix.tv2}
\alias{.dzetaIJ.matrix}
\alias{.dzetaKappa.intercept}
\alias{.dzetaKappa.matrix}
\alias{.dzetaKappa.matrix.simp.t}
\alias{.dzetaKappa.matrix2}
\alias{.dzetaKappa.simp.intercept}
\alias{.EvaluatePoint}
\alias{.generategrid}
\alias{.generatenames}
\alias{.givens.mult.M}
\alias{.givens.real}
\alias{.GridEvaluatePointA}
\alias{.interpret.intercept}
\alias{.inv.rho.givens.real}
\alias{.Kappa.matrix}
\alias{.Kappa.matrix.NA}
\alias{.M.mult.givens}
\alias{.min.mLogLik.BmPsi}
\alias{.min.RSS.BmPsi}
\alias{.MinusPhylLogLikFunc}
\alias{.MinusPhylLogLikFuncMVslouch}
\alias{.MinusPhylLogLikFuncouch}
\alias{.num.inv.transform.positive}
\alias{.num.transform.positive}
\alias{.ouch.design.matrix}
\alias{.ouch.GLS}
\alias{.par.inv.transform}
\alias{.par.inv.transform.decomp.matrix}
\alias{.par.inv.transform.decomp.neg.real.matrix}
\alias{.par.inv.transform.decomp.pos.real.matrix}
\alias{.par.inv.transform.decomp.real.matrix}
\alias{.par.inv.transform.decomp.sym.matrix}
\alias{.par.inv.transform.invert.matrix}
\alias{.par.inv.transform.invert.qr.matrix}
\alias{.par.inv.transform.invert.svd.matrix}
\alias{.par.inv.transform.lowertri.matrix}
\alias{.par.inv.transform.normal.matrix}
\alias{.par.inv.transform.orth.matrix.cayley}
\alias{.par.inv.transform.orth.matrix.givens}
\alias{.par.inv.transform.symmetric.matrix}
\alias{.par.inv.transform.twobytwo.matrix}
\alias{.par.inv.transform.uppertri.matrix}
\alias{.par.transform}
\alias{.par.transform.decomp.matrix}
\alias{.par.transform.decomp.neg.real.matrix}
\alias{.par.transform.decomp.pos.real.matrix}
\alias{.par.transform.decomp.real.matrix}
\alias{.par.transform.decomp.sym.matrix}
\alias{.par.transform.invert.matrix}
\alias{.par.transform.invert.qr.matrix}
\alias{.par.transform.invert.svd.matrix}
\alias{.par.transform.lowertri.matrix}
\alias{.par.transform.normal.matrix}
\alias{.par.transform.orth.matrix.cayley}
\alias{.par.transform.orth.matrix.givens}
\alias{.par.transform.symmetric.matrix}
\alias{.par.transform.twobytwo.matrix}
\alias{.par.transform.uppertri.matrix}
\alias{.params.summary.bm}
\alias{.params.summary.mvslouch}
\alias{.params.summary.ouch}
\alias{.resizegrid}
\alias{.row.rot}
\alias{.set.estimparams}
\alias{.set.maxiter}
\alias{.set.paramatrizationnames}
\alias{.set.tol}
\alias{.slouch.mv.design.matrix.YcT}
\alias{.slouch.mv.iGLS}
\alias{.slouch.mv.residual.covar}
\alias{.ouch2slouch.mvsl}
\alias{.solve.reg}
\alias{.make.states.mvsl}
\description{Internal mvSLOUCH objects.}
\details{These are not to be called by the user.}
\keyword{internal}