\name{mvSLOUCH-internal}
\title{Internal mvSLOUCH objects}
\alias{.ape2slouch.mvsl}
\alias{.beginEstimationParams}
\alias{.bm.estim}                                
\alias{.bm.phyl.cov}                             
\alias{.bm.sim.est}                              
\alias{.bm.simulate}                             
\alias{.boot.cleanup}                            
\alias{.boot.extract}                            
\alias{.boot.extract.model.components}           
\alias{.boot.getval}                             
\alias{.calc_species_covariance}                 
\alias{.calc_species_mean}                       
\alias{.calc.cov.bm.mv}                          
\alias{.calc.cov.ouch.mv}                        
\alias{.calc.cov.slouch.mv}                      
\alias{.calc.exptA}                              
\alias{.calc.exptA.2dim}                         
\alias{.calc.integral.evAStevA}                  
\alias{.calc.integral.evAStevA.2dim}             
\alias{.calc.mean.bm.mv}                         
\alias{.calc.mean.ouch.mv}                       
\alias{.calc.mean.slouch.mv}                     
\alias{.calc.phyl.cov}                           
\alias{.calc.phyl.halflife}                      
\alias{.calc.phyl.LogLik.traits}                 
\alias{.calc.vec.dist}                           
\alias{.calcCI}                                  
\alias{.calcESSanalytical}                       
\alias{.calcESSsim}                              
\alias{.calculate_Tree_dists_forCovariance}      
\alias{.CalcVlq}                                 
\alias{.CalcVlq2}                                
\alias{.CalcVlqStat}                             
\alias{.callPCMBase_mvlik}                       
\alias{.changeSigmatoSyy}                        
\alias{.CheckSanityRegimes}                      
\alias{.cleanUpModelParams}                      
\alias{.col.rot}                                 
\alias{.correct_npar0}                           
\alias{.correct_phylGLS_response_by_intercept}   
\alias{.correct.names}                           
\alias{.createMeasurementError}                  
\alias{.createRegimes}                           
\alias{.createStartPointsASyyB}                  
\alias{.decompEigenA.S}                          
\alias{.describe.best.model}                     
\alias{.design_matrix_bm}                        
\alias{.design_matrix_construction}              
\alias{.design_matrix_mvslouch}                  
\alias{.design_matrix_ouch}                      
\alias{.detV}                                    
\alias{.do_phylGLSestimation}                    
\alias{.draw.bm}                                 
\alias{.draw.mvslouch}                           
\alias{.draw.ouch}                               
\alias{.drawOneLineage}                          
\alias{.dzeta.matrix.t}                          
\alias{.dzeta.matrix.t1t2}                       
\alias{.dzeta.matrix.t1t2v2}                     
\alias{.dzeta.matrix.tv2}                        
\alias{.dzetaIJ.matrix}                          
\alias{.dzetaKappa.intercept}                    
\alias{.dzetaKappa.matrix}                       
\alias{.dzetaKappa.matrix.simp.t}                
\alias{.dzetaKappa.matrix2}                      
\alias{.dzetaKappa.simp.intercept}               
\alias{.EvaluatePoint}                           
\alias{.extract_from_signs}                     
\alias{.extract_GLS_results}                     
\alias{.extract_GLS_results_bm}                  
\alias{.extract_GLS_results_mvslouch}            
\alias{.extract_GLS_results_ouch}                
\alias{.extract.model.key.properties}            
\alias{.generate.all.model.setups}               
\alias{.generate.basic.model.setups}             
\alias{.generate.ext.model.setups}               
\alias{.generate.fund.model.setups}              
\alias{.generate.list.of.model.setups}           
\alias{.generate.model.description}              
\alias{.generate.model.se}                       
\alias{.generate.univ.model.setups}              
\alias{.generatenames}                           
\alias{.getESScriteria}                          
\alias{.getMVSLphylCovMatrix}                    
\alias{.getMVSLproc_size}                        
\alias{.givens.mult.M}                           
\alias{.givens.real}                             
\alias{.glsgc.estim}                             
\alias{.InitialRegimeSetup}                      
\alias{.internal_BrownianMotionModel}            
\alias{.internal_mvslouchModel}                  
\alias{.internal_ouchModel}                      
\alias{.inv.rho.givens.real}                     
\alias{.Kappa.matrix}                            
\alias{.Kappa.matrix.NA}                         
\alias{.M.mult.givens}                           
\alias{.make.states.mvsl}                        
\alias{.maxlik.estim}                            
\alias{.MinusPhylLogLikFunc}                     
\alias{.MinusPhylLogLikFuncMVslouch}             
\alias{.MinusPhylLogLikFuncouch}                 
\alias{.ModelSetupPCMBase}                       
\alias{.mvslouch_to_ouch_model}                  
\alias{.mvslouch.phyl.cov}                       
\alias{.mvslouch.simulate}                       
\alias{.norm.max}                                
\alias{.num.inv.transform.positive}              
\alias{.num.transform.positive}                  
\alias{.oubm.sim.est}                            
\alias{.ouch.phyl.cov}                           
\alias{.ouch.simulate}                           
\alias{.ouch2slouch.mvsl}                        
\alias{.ouou.sim.est}                            
\alias{.OUtoBM_pcmbase_model_box}                
\alias{.packageName}                             
\alias{.par_transform_withPCMBase}               
\alias{.par.inv.transform}                       
\alias{.par.inv.transform.decomp.matrix}         
\alias{.par.inv.transform.decomp.neg.real.matrix}
\alias{.par.inv.transform.decomp.pos.real.matrix}
\alias{.par.inv.transform.decomp.real.matrix}    
\alias{.par.inv.transform.decomp.sym.matrix}     
\alias{.par.inv.transform.invert.matrix}         
\alias{.par.inv.transform.invert.qr.matrix}      
\alias{.par.inv.transform.invert.svd.matrix}     
\alias{.par.inv.transform.lowertri.matrix}       
\alias{par.inv.transform.orth.matrix.cayley}    
\alias{par.inv.transform.orth.matrix.givens}    
\alias{.par.inv.transform.symmetric.matrix}      
\alias{.par.inv.transform.twobytwo.matrix}       
\alias{.par.inv.transform.uppertri.matrix}       
\alias{.par.transform}                          
\alias{.par.transform.decomp.matrix}             
\alias{.par.transform.decomp.neg.real.matrix}    
\alias{.par.transform.decomp.pos.real.matrix}    
\alias{.par.transform.decomp.real.matrix}        
\alias{.par.transform.decomp.sym.matrix}         
\alias{.par.transform.invert.matrix}             
\alias{.par.transform.invert.qr.matrix}          
\alias{.par.transform.invert.svd.matrix}         
\alias{.par.transform.lowertri.matrix}           
\alias{.par.transform.orth.matrix.cayley}        
\alias{.par.transform.orth.matrix.givens}        
\alias{.par.transform.symmetric.matrix}          
\alias{.par.transform.twobytwo.matrix}           
\alias{.par.transform.uppertri.matrix}           
\alias{.Params.summary}                          
\alias{.params.summary.bm}                       
\alias{.params.summary.mvslouch}                 
\alias{.params.summary.ouch}                     
\alias{.pcmbaseDphylGaussian_RSS}                
\alias{.pcmbaseDphylOU_GLS}                      
\alias{.PhyloSDEestim}                           
\alias{.phyltree_remove_path_fields}
\alias{.row.rot}
\alias{.set_mean0_pcmbase_model_box}
\alias{.set.estimparams}
\alias{.set.maxiter}
\alias{.set.paramatrizationnames}
\alias{.set.tol}
\alias{.simulVasicekProcPhylTree}
\alias{.SummarizeFullPoint}
\alias{.sym.par}
\alias{.sym.unpar}
\alias{.update_pcmbase_box_params}
\alias{.update_pcmbase_box_params_bm}
\alias{.update_pcmbase_box_params_mvslouch}
\alias{.update_pcmbase_box_params_ouch}
\alias{.internal_phyltree_paths_BL}
\alias{.check_input_trait_data}
\alias{.is_det0}
\alias{.my_cov2cor}
\alias{.my_stop}
\alias{.my_message}
\alias{.set_sing_bl_threshold}
\description{Internal mvSLOUCH objects.}
\details{These are not to be called by the user.}
\keyword{internal}
